Removed duplicated modules
https://github.com/fedify-dev/fedify/pull/451#discussion_r2403733075
