+++ packages/fedify/src/nodeinfo/client.ts
@@ -1,5 +1,5 @@
+import { getUserAgent, type GetUserAgentOptions } from "@fedify/vocab-runtime";
 import { getLogger } from "@logtape/logtape";
-import { getUserAgent, type GetUserAgentOptions } from "../utils/request.ts";
 import type { ResourceDescriptor } from "../webfinger/jrd.ts";
 import type {
   InboundService,
