+++ packages/fedify/src/utils/docloader.test.ts
@@ -1,3 +1,4 @@
+import { UrlError } from "@fedify/vocab-runtime";
 import { assertEquals, assertRejects } from "@std/assert";
 import fetchMock from "fetch-mock";
 import { verifyRequest } from "../sig/http.ts";
@@ -5,7 +6,6 @@ import { mockDocumentLoader } from "../testing/docloader.ts";
 import { rsaPrivateKey2 } from "../testing/keys.ts";
 import { test } from "../testing/mod.ts";
 import { getAuthenticatedDocumentLoader } from "./docloader.ts";
-import { UrlError } from "./url.ts";
 
 test("getAuthenticatedDocumentLoader()", async (t) => {
   fetchMock.spyGlobal();
