+++ packages/fedify/src/utils/mod.ts
@@ -1,3 +1,2 @@
 export { getAuthenticatedDocumentLoader } from "./docloader.ts";
 export { kvCache } from "./kv-cache.ts";
-export { getUserAgent } from "./request.ts";
