+++ packages/fedify/src/vocab/actor.ts
@@ -1,7 +1,7 @@
+import type { GetUserAgentOptions } from "@fedify/vocab-runtime";
 import { SpanStatusCode, trace, type TracerProvider } from "@opentelemetry/api";
 import { domainToASCII, domainToUnicode } from "node:url";
 import metadata from "../../deno.json" with { type: "json" };
-import type { GetUserAgentOptions } from "../utils/request.ts";
 import { lookupWebFinger } from "../webfinger/lookup.ts";
 import { getTypeId } from "./type.ts";
 import { Application, Group, Organization, Person, Service } from "./vocab.ts";
